Refactor OAS3 builders to reduce complexity and class length
Split large classes/modules into smaller, focused units to address
rubocop Metrics offenses:

**oas30.rb (357→326 lines)**
- Extract SchemaExporter module with schema export methods
- Extract SchemaFields module with field helper methods

**schema_builder.rb**
- Extract build_from_definition into smaller helper methods
- Reduce cyclomatic complexity from 19 to under limit

**from_routes.rb (754→448 lines)**
- Extract ParameterBuilder module for parameter handling
- Extract ParamSchemaBuilder module for schema building
- Extract RequestBodyBuilder module for request body handling
- Extract ResponseBuilder module for response building

All 771 tests pass (293 OAS3, 478 Swagger 2.0).
No rubocop offenses in openapi/builder and exporter directories.
